Move apply_manual_codes into project_reach as a library file
