--- apply_manual_codes/Pipfile
@@ -1,12 +1,0 @@
-[[source]]
-url = "https://pypi.python.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-CoreDataModules = {editable = true, ref = "v0.7.0", git = "https://www.github.com/AfricasVoices/CoreDataModules"}
-
-[dev-packages]
-
-[requires]
-python_version = "3.6"
